Move TemporaryFence/TemporaryCommandBuffer out of dump-resources into vulkan_temporary_objects.h/.cpp
- add CreateAndBegin(queue_family_index) overload
- use GetDeviceQueue2-aware queue retrieval, check EndCommandBuffer/QueueSubmit results
- use TemporaryCommandBuffer in ProcessSetSwapchainImageStateCommand
