+++ framework/decode/vulkan_replay_dump_resources_common.cpp
@@ -1395,77 +1395,6 @@ void CullDescriptors(const CommonObjectInfoTable&             object_info_table_
     }
 }
 
-VkResult TemporaryCommandBuffer::CreateAndBegin(graphics::FindQueueFamilyIndex_fp queue_finder_fp)
-{
-    const uint32_t queue_index = queue_finder_fp(device_info.enabled_queue_family_flags);
-    GFXRECON_ASSERT(queue_index != VK_QUEUE_FAMILY_IGNORED);
-
-    const VkCommandPoolCreateInfo pool_create_info = { VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO,
-                                                       nullptr,
-                                                       VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT,
-                                                       queue_index };
-    VkResult res = device_table.CreateCommandPool(device_info.handle, &pool_create_info, nullptr, &command_pool);
-    if (res != VK_SUCCESS)
-    {
-        GFXRECON_LOG_ERROR("%s() CreateCommandPool failed (%s)", __func__, util::ToString(res).c_str());
-        return res;
-    }
-
-    const VkCommandBufferAllocateInfo alloc_info = {
-        VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO, nullptr, command_pool, VK_COMMAND_BUFFER_LEVEL_PRIMARY, 1
-    };
-    res = device_table.AllocateCommandBuffers(device_info.handle, &alloc_info, &command_buffer);
-    if (res != VK_SUCCESS)
-    {
-        GFXRECON_LOG_ERROR("%s() AllocateCommandBuffers failed (%s)", __func__, util::ToString(res).c_str());
-        return res;
-    }
-
-    device_table.GetDeviceQueue(device_info.handle, queue_index, 0, &queue);
-
-    device_table.ResetCommandBuffer(command_buffer, VkCommandBufferResetFlagBits(0));
-
-    const VkCommandBufferBeginInfo begin_info = {
-        VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO, nullptr, VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT, nullptr
-    };
-
-    res = device_table.BeginCommandBuffer(command_buffer, &begin_info);
-    if (res != VK_SUCCESS)
-    {
-        GFXRECON_LOG_ERROR("%s() BeginCommandBuffer failed (%s)", __func__, util::ToString(res).c_str());
-        return res;
-    }
-
-    return VK_SUCCESS;
-}
-
-VkResult TemporaryCommandBuffer::SubmitAndDestroy()
-{
-    GFXRECON_ASSERT(command_buffer != VK_NULL_HANDLE);
-    GFXRECON_ASSERT(queue != VK_NULL_HANDLE);
-    GFXRECON_ASSERT(command_pool != VK_NULL_HANDLE);
-
-    TemporaryFence fence(device_info.handle, device_table);
-
-    device_table.EndCommandBuffer(command_buffer);
-
-    const VkSubmitInfo submit_info = {
-        VK_STRUCTURE_TYPE_SUBMIT_INFO, nullptr, 0, nullptr, nullptr, 1, &command_buffer, 0, nullptr
-    };
-    device_table.QueueSubmit(queue, 1, &submit_info, fence.handle);
-
-    VkResult res = fence.Wait();
-    if (res != VK_SUCCESS)
-    {
-        return res;
-    }
-
-    device_table.DestroyCommandPool(device_info.handle, command_pool, nullptr);
-    command_pool = VK_NULL_HANDLE;
-
-    return VK_SUCCESS;
-}
-
 static VkResult SerializeAccelerationStructure(AccelerationStructureDumpResourcesContext* acceleration_structure,
                                                const VulkanDeviceInfo*                    device_info,
                                                const graphics::VulkanDeviceTable&         device_table,
